vmware: set the driver version
Since commit 7457cbe8 the vmware driver version isn't set anymore.

+++ src/vmware/vmware_conf.h
@@ -48,7 +48,7 @@ struct vmware_driver {
     virDomainXMLOptionPtr xmlopt;
 
     virDomainObjListPtr domains;
-    int version;
+    unsigned long version;
     int type;
     char *vmrun;
 };
